io: move WriteHasher from osmet/unpacker
Those bits are currently only used in osmet, but soon won't be. There's
nothing osmet-specific about the code otherwise.
